app: move POST /api/client to app.routers.client
